Bump the version before announcing
We were erroneously sending new version announcements off to Discord
before we actually bumped the version number in the product object.
This caused the announcement to say the old version number and not the
new version number.

This commit changes where the version number is incremented so that the
announcement reflects the actual new version number.